+++ agent/src/main/java/com/teamscale/jacoco/agent/AgentOptionsParser.java
@@ -156,6 +156,8 @@ public class AgentOptionsParser {
 				return true;
 			case "ignore-duplicates":
 				options.shouldIgnoreDuplicateClassFiles = Boolean.parseBoolean(value);
+			case "dump-on-exit":
+				options.shouldDumpOnExit = Boolean.parseBoolean(value);
 				return true;
 			case "includes":
 				options.jacocoIncludes = value.replaceAll(";", ":");